cf_delete_old_images/requirements.txt
google-cloud-artifact-registry
google-auth
